Push notification added

+++ client/src/app/push-notifications/push-notifications.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Output, Input, EventEmitter } from '@angular/core';
 import { SwPush } from '@angular/service-worker';
 import { Observable } from 'rxjs';
+import { take } from 'rxjs/operators';
 import { DeviceDetectorService } from 'ngx-device-detector';
 
 import { ConfigService } from '../config.service';
@@ -21,6 +22,7 @@ export class PushNotificationsComponent implements OnInit {
   pushIsEnabled: boolean;
   isPushSubscription: boolean;
   isSafari: boolean;
+  
   @Output() subscriptionError: EventEmitter<any> = new EventEmitter();
 
   ngOnInit() {
@@ -38,47 +40,59 @@ export class PushNotificationsComponent implements OnInit {
   onChange($event: any): void {
   	if($event) {
   		this.subscribeToPush();
-  	} else this.unsubscribeFromPush();
+  	} else {
+  		this.unsubscribeFromPush();
+  	}
   }
 
 
-  subscribeToPush() {
+  private subscribeToPush() {
     // Requesting messaging service to subscribe current client (browser)
     this.swPush.requestSubscription({
       serverPublicKey: this.conf.vapidPublicKey
     })
       .then(pushSubscription => {
-      	// Passing subscription object to app server
+      		// Passing subscription object to app server
       	this.userService.sendSubscription(pushSubscription)
       		.subscribe(res => {}, error => { 
-		  		this.subscriptionError.emit('Error occured while subscribing to push notifications');
+		  		this.subscriptionError
+		  			.emit('Error occured while subscribing to push notifications');
 		  	});
       })
       .catch(err => {
-	  		this.subscriptionError.emit('Error occured while subscribing to push notifications');
+	  		this.subscriptionError
+	  			.emit('Error occured while subscribing to push notifications');
       });
   }
 
 
-  unsubscribeFromPush() {
-  	this.swPush.subscription
-  		.subscribe(pushSubscription => {
-  			this.userService.deleteSubscription(pushSubscription.endpoint)
-  				.subscribe(() => {
-  					pushSubscription.unsubscribe()
-              .then(() => {
-                this.isPushSubscription = false;
-              })
-              .catch(err => {
-                this.subscriptionError
-                	this.isPushSubscription = true;
-                	this.subscriptionError.emit('Error occured while unsubscribing from push notification');
-              });
-  				},
-  				error => { 
-  					this.isPushSubscription = true;
-			  		this.subscriptionError.emit('Error occured while subscribing to push notifications');
-			  	});
+  private unsubscribeFromPush() {
+  	this.swPush.subscription.pipe(take(1)).subscribe(pushSubscription => {
+  		if(!pushSubscription) {
+  			this.subscriptionError
+      		.emit('Error occured while unsubscribing from push notification');
+      		return;
+  		}
+
+			this.userService.deleteSubscription(pushSubscription.endpoint)
+				.subscribe(() => {
+					pushSubscription.unsubscribe()
+            .then(() => {
+              // this.isPushSubscription = false;
+            })
+            .catch(err => {
+              // this.subscriptionError
+            	// this.isPushSubscription = true;
+            	this.subscriptionError
+            		.emit('Error occured while unsubscribing from push notification');
+            });
+				},
+				error => { 
+					console.log('Server error');
+					// this.isPushSubscription = true;
+		  		this.subscriptionError
+		  			.emit('Error occured while unsubscribing from push notification');
+		  	});
   		});
   }
 }
